+++ samples/rag-aisearch/csharp-ooproc/FilePrompt.cs
@@ -18,9 +18,6 @@ public static class FilePrompt
     {
         [JsonPropertyName("Url")]
         public string? Url { get; set; }
-
-        [JsonPropertyName("Title")]
-        public string? Title { get; set; }
     }
 
     public class SemanticSearchRequest
@@ -29,26 +26,38 @@ public static class FilePrompt
         public string? Prompt { get; set; }
     }
 
-    // REVIEW: There are several assumptions about how the Embeddings binding and the SemanticSearch bindings
-    //         work together. We should consider creating a higher-level of abstraction for this.
     [Function("IngestFile")]
-    public static async Task<HttpResponseData> IngestFile(
-        [HttpTrigger(AuthorizationLevel.Function, "post")] HttpRequestData req,
-        [EmbeddingsStoreInput("{Url}", InputType.Url, "{Title}", "AISearchEndpoint", "openai-index", Model = "%EMBEDDING_MODEL_DEPLOYMENT_NAME%")] EmbeddingsContext embeddings)
+    public static async Task<EmbeddingsStoreOutputResponse> IngestFile(
+        [HttpTrigger(AuthorizationLevel.Function, "post")] HttpRequestData req)
     {
         using StreamReader reader = new(req.Body);
         string request = await reader.ReadToEndAsync();
 
         EmbeddingsRequest? requestBody = JsonSerializer.Deserialize<EmbeddingsRequest>(request);
 
-        if (requestBody == null || requestBody.Url == null || requestBody.Title == null)
+        if (requestBody == null || requestBody.Url == null)
         {
-            throw new ArgumentException("Invalid request body. Make sure that you pass in {\"Url\": value , \"Title\": value} as the request body.");
+            throw new ArgumentException("Invalid request body. Make sure that you pass in {\"Url\": value } as the request body.");
         }
 
-        HttpResponseData response = req.CreateResponse(HttpStatusCode.OK);
-        await response.WriteAsJsonAsync(new { status = "success", requestBody.Title, chunks = embeddings.Count });
-        return response;
+        Uri uri = new(requestBody.Url);
+        string filename = Path.GetFileName(uri.AbsolutePath);
+
+        HttpResponseData response = req.CreateResponse(HttpStatusCode.Created);
+
+        return new EmbeddingsStoreOutputResponse
+        {
+            HttpResponse = response,
+            SearchableDocument = new SearchableDocument(filename)
+        };
+    }
+
+    public class EmbeddingsStoreOutputResponse
+    {
+        [EmbeddingsStoreOutput("{url}", InputType.Url, "AISearchEndpoint", "openai-index", Model = "%EMBEDDING_MODEL_DEPLOYMENT_NAME%")]
+        public SearchableDocument? SearchableDocument { get; set; }
+
+        public HttpResponseData? HttpResponse { get; set; }
     }
 
     [Function("PromptFile")]