+++ src/WebJobs.Extensions.OpenAI.AzureAISearch/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Added support for key authentication back (use of managed identity is always recommended)
+- Updated Microsoft.Azure.WebJobs.Extensions.OpenAI to 0.15.0
 
 ## v0.2.0 - 2024/04/24