+++ src/WebJobs.Extensions.OpenAI.Kusto/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.15.0 - Unreleased
+
+### Changed
+
+- Updated Microsoft.Azure.WebJobs.Extensions.OpenAI to 0.15.0
+
 ## v0.14.0 - 2024/04/24
 
 ### Changed